% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_IDDS.R
\name{assign_IDDS}
\alias{assign_IDDS}
\title{Assign a Unique Sample ID to Each Row of a Data Frame (Assign Function)}
\usage{
assign_IDDS(object, id_name)
}
\arguments{
\item{object}{character. The name of the data frame on the server to which the IDs will be assigned.}

\item{id_name}{character. The name of the new column to store the unique IDs.}
}
\value{
A modified data frame with a new column containing unique IDs. The object is overwritten server-side.
}
\description{
This server-side assign function generates a column of unique alphanumeric sample IDs and assigns them to each row of the specified data frame.
}
\details{
This function is useful for generating unique row identifiers in federated settings where row-level tracking is needed but data must remain anonymized.
}
\author{
Roy Gusinow
}
