% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersionDS.R
\name{dispersionDS}
\alias{dispersionDS}
\title{aggregate function -- create dispersion for clustered error estimate}
\usage{
dispersionDS(formula, data, pred_col, family)
}
\arguments{
\item{formula}{character string formulaula (e.g., "y ~ x1 + x2")}

\item{data}{character string name of the dataframe}

\item{pred_col}{character string name of column with predictions}

\item{family}{character GLM family (e.g., "poisson", "binomial", "gaussian")}
}
\value{
numeric dispersion value (scalar)
}
\description{
Computes a dispersion estimate using weighted residuals from a regression model.
For GLM families like Poisson or Binomial, the dispersion is set to 1. For others, it computes
the sum of squared residuals.
}
\details{
Estimate dispersion from weighted residuals (Aggregate function)

Aggregate function that returns a non-disclosive dispersion measure. A disclosure control check is applied to ensure that the number of residuals exceeds `nfilter.tab`.
}
\author{
Roy Gusinow
}
