% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCDFDS.R
\name{eCDFDS}
\alias{eCDFDS}
\title{aggregate function -- empirical cumulative distribution function}
\usage{
eCDFDS(object, min, max, len, weights)
}
\arguments{
\item{object}{character string specifying the name of the vector on the server}

\item{min}{numeric minimum of the domain}

\item{max}{numeric maximum of the domain}

\item{len}{integer length of the output eCDF vector (number of points in domain)}

\item{weights}{optional; currently unused}
}
\value{
A numeric vector of cumulative probabilities at each point in the domain
}
\description{
Computes the eCDF over a user-defined domain range and resolution, avoiding disclosure by capping the resolution and verifying data length.
}
\details{
eCDF vector for a numeric variable

This is a server-side aggregate function for drawing a non-disclosive empirical CDF. Resolution is limited to reduce disclosure risk. Can be paired with a client function that plots this result.
}
\author{
Roy Gusinow
}
