% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchitDS.R
\name{matchitDS}
\alias{matchitDS}
\title{assign function -- new dataframe with matched records}
\usage{
matchitDS(formula, data, distance, id_name)
}
\arguments{
\item{formula}{formula, character string specifying the treatment variable}

\item{data}{character string name of the dataframe}

\item{distance}{character string name of column in data containing distance measure}

\item{id_name}{character string for name of ID column in data}
}
\value{
A dataframe with ID, treatment, and noise-injected distance column
}
\description{
Applies a privacy-preserving transformation to propensity score distances for federated matching. Uses noise injection (normal or k-NN) and ensures safe thresholds per DataSHIELD disclosure settings.
}
\details{
Matchit pooled

Disclosure control includes minimum unique values and safe thresholds for standard deviation and k.
}
\author{
Roy Gusinow
}
