% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchitDS2.R
\name{matchitDS2}
\alias{matchitDS2}
\title{assign function -- new dataframe with matched records}
\usage{
matchitDS2(data, ids, distances, weights, id_name)
}
\arguments{
\item{data}{character string name of the dataframe}

\item{ids}{vector of matched IDs}

\item{distances}{vector of distances for matched records}

\item{weights}{vector of weights for matched records}

\item{id_name}{character string, name of the ID column in `data`}
}
\value{
Subsetted dataframe with matched records and distance/weights columns filled
}
\description{
Integrates matched ID records with distance and weights. Returns subset of matched individuals only.
}
\details{
Matchit postprocessing (pooled)

Applies disclosure checks to prevent returning too few matched records (using `nfilter.tab`)
}
\author{
Roy Gusinow
}
